Extract SVG code to a separate module
