+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "hangman",
-  "version": "0.0.15",
+  "version": "0.0.16",
   "private": true,
   "proxy": "http://localhost:3001/",
   "dependencies": {
