+++ src/constants.js
@@ -1,3 +1,5 @@
+import { NUM_GALLOWS_STEPS } from './Components/gallowsSvg';
+
 export const GAME_STATE_CHOOSING_WORD = 0;
 export const GAME_STATE_IN_PROGRESS = 1;
 export const GAME_STATE_GAME_OVER = 2;
@@ -6,4 +8,4 @@ export const OUTCOME_NONE = 0;
 export const OUTCOME_WON = 1;
 export const OUTCOME_LOST = 2;
 
-export const MAX_BAD_GUESSES = 11;
+export const MAX_BAD_GUESSES = NUM_GALLOWS_STEPS;
